+++ src/lib/libast/misc/cmdlib.h
@@ -2,7 +2,7 @@
 *                                                                      *
 *               This software is part of the ast package               *
 *          Copyright (c) 1985-2012 AT&T Intellectual Property          *
-*          Copyright (c) 2020-2022 Contributors to ksh 93u+m           *
+*          Copyright (c) 2020-2026 Contributors to ksh 93u+m           *
 *                      and is licensed under the                       *
 *                 Eclipse Public License, Version 2.0                  *
 *                                                                      *
@@ -14,6 +14,7 @@
 *                  David Korn <dgk@research.att.com>                   *
 *                   Phong Vo <kpv@research.att.com>                    *
 *                  Martijn Dekker <martijn@inlv.org>                   *
+*            Johnothan King <johnothanking@protonmail.com>             *
 *                                                                      *
 ***********************************************************************/
 /*
@@ -34,9 +35,9 @@
 	int		argcount;	/* current arg count		*/ \
 	int		argmax;		/* max # args			*/ \
 	int		echo;		/* just an echo			*/ \
-	int		flags;		/* CMD_* flags			*/ \
-	int		insertlen;	/* strlen(insert)		*/ \
-	int		offset;		/* post arg offset		*/ \
+	uint32_t	flags;		/* CMD_* flags			*/ \
+	size_t		insertlen;	/* strlen(insert)		*/ \
+	ptrdiff_t	offset;		/* post arg offset		*/ \
 	Cmddisc_t*	disc;		/* discipline			*/ \
 	char**		argv;		/* exec argv			*/ \
 	char**		firstarg;	/* first argv file arg		*/ \
